Fixed not being able to auto translate with google
Reverted changes, reasons:
It broke the auto translation feature, which is by far the most important one.
It was way too heavy and slow. Many images and files.
It was way taking too much visual space and not a good ux.

To not lose the good parts of the work done, I tried to keep it simple and mimic that visual, using what worked best.
Hopefully it still look as good as the design changes made, while lightweight and easy ux.

Future goals:
Improve the auto translate hack into a proper (and reliable) tool.
Add instructions / user guides.